Decided to use f32 for possibly better performance
We don't need high precision

+++ src/math_utils/matrix.rs
@@ -2,19 +2,19 @@ use super::vector::{Vector3, Vector4};
 
 
 pub struct Matrix3 {
-    x: Vector3<f64>,
-    y: Vector3<f64>,
-    z: Vector3<f64>
+    x: Vector3<f32>,
+    y: Vector3<f32>,
+    z: Vector3<f32>
 }
 
 pub struct Matrix4 {
-    x: Vector4<f64>,
-    y: Vector4<f64>,
-    z: Vector4<f64>,
-    w: Vector4<f64>
+    x: Vector4<f32>,
+    y: Vector4<f32>,
+    z: Vector4<f32>,
+    w: Vector4<f32>
 }
 
 pub struct TransformMatrix {
     matrix: Matrix3,
-    translation: Vector3<f64>
+    translation: Vector3<f32>
 }
